+++ djstripe/fields.py
@@ -49,8 +49,9 @@ class StripeCurrencyCodeField(models.CharField):
 class StripeQuantumCurrencyAmountField(models.BigIntegerField):
     """
     A field used to store currency amounts in cents (etc) as per stripe.
+    By contacting stripe support, some accounts will have their limit raised to 11
+    digits, hence the use of BigIntegerField instead of IntegerField
     """
-
     pass
 
 
@@ -66,7 +67,10 @@ class StripeDecimalCurrencyAmountField(models.DecimalField):
     """
 
     def __init__(self, *args, **kwargs):
-        """Assign default args to this field."""
+        """
+        Assign default args to this field. By contacting stripe support, some accounts
+        will have their limit raised to 11 digits
+        """
         defaults = {"decimal_places": 2, "max_digits": 11}
         defaults.update(kwargs)
         super().__init__(*args, **defaults)